+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 9.2.1
 - SQL set default nextval (auto increment) for RosarioSIS version < 5.0 on install & old add-ons in Update.fnc.php
 - SQL no more cast MARKING_PERIOD_ID column as text/varchar in rosariosis.sql & InputFinalGrades.php
 - PLpgSQL compact & consistent function declaration in rosariosis.sql
+- Use DB transaction statements compatible with MySQL in database.inc.php
 
 Changes in 9.2
 --------------